changed millis() variables to unsigned long

+++ examples/mitsubishi_heatpump_mqtt_esp8266/mitsubishi_heatpump_mqtt_esp8266.ino
@@ -15,7 +15,7 @@
 WiFiClient espClient;
 PubSubClient mqtt_client(espClient);
 HeatPump hp;
-uint lastTempSend;
+unsigned long lastTempSend;
 
 // debug mode, when true, will send all packets received from the heatpump to topic heatpump_debug_topic
 // this can also be set by sending "on" to heatpump_debug_set_topic
